# docs/openapi/aether-v0.yaml
openapi: 3.1.0
info:
  title: Aether /api/v0 API
  version: 0.1.0
  contact:
    name: Andres Bott
    url: https://github.com/andresbott/aether
  description: >
    Internal, admin-only server-management API — collections/libraries,
    scanning, tasks/scheduling, and other administration concerns. This is
    NOT the music-client API: browsing, playback, playlists, album/artist
    navigation and search live exclusively under `/rest`, which must stay
    OpenSubsonic-compliant and has no OpenAPI document of its own (a Subsonic
    client does not consume one).


    **Scope.** This document describes the full `/api/v0` surface as mounted
    in `app/router/api_v0.go`: the metadata editor (pictures, raw tags,
    folder/track browsing, identify and identify-album, artist images),
    artists, users, tasks, tokens, radiobrowser, auth, libraries, and
    health/version/me.
servers:
  - url: /api/v0
    description: >
      Every path below is mount-relative. The base is modelled exclusively
      through this `servers` entry, so bumping the version prefix — this is
      `/api/v0` while the API is pre-1.0 — stays a one-line change, with no
      path or operationId edits.
...
security:
  - sessionCookie: []
paths:
...
  /libraries:
    get:
      operationId: listLibraries
      summary: List every configured collection/library.
      description: >
        Admin. Includes both API-managed ("db") and config-file-provisioned
        ("config") libraries — the two sources are additive, not exclusive,
        so the UI can render config-provisioned rows locked (no edit/delete)
        alongside editable ones. Unpaginated, like listUsers.
      tags: [libraries]
      responses:
        '200':
          description: Every configured library.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LibraryList' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '500': { $ref: '#/components/responses/InternalServerError' }
    post:
      operationId: createLibrary
      summary: Create a library owned by this API (source "db").
      description: >
        Admin. `name` and `path` are required: a missing value answers 400
        (validation_error); a present-but-invalid one (name too long, or a
        path that isn't an existing, readable directory) answers 422,
        itemising `/name` or `/path`. `path` is resolved to an absolute
        path server-side before being stored. `exclude_patterns`,
        `default_view`, `icon` and `cover_style` are always optional, so
        any failure on them (a regex that fails to compile, an unknown
        view/style, an unusable icon name) is unconditionally 422 — never
        400 — itemising the failing pointer.

        Refused with 409 `config_managed` when the submitted `name` or
        `path` is already owned by a config-provisioned library, distinct
        from the plain 409 a duplicate against a db-owned library also
        answers with (both share the `Conflict` response's generic Problem
        shape).
      tags: [libraries]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LibraryCreateRequest' }
      responses:
        '201':
          description: Created.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Library' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
        '500': { $ref: '#/components/responses/InternalServerError' }
  /libraries/{id}:
    get:
      operationId: getLibrary
      summary: Fetch one library.
      description: >
        Admin. `track_count` is computed fresh on every read (a live query
        against the tracks table), not a cached column.
      tags: [libraries]
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200':
          description: The library.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Library' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalServerError' }
    put:
      operationId: updateLibrary
      summary: Replace a library's editable fields.
      description: >
        Admin. `name`/`path`/`exclude_patterns`/`default_view`/`icon`/
        `cover_style` are always fully replaced by the submitted value
        (empty/omitted `default_view`, `icon` and `cover_style` fall back
        the same way createLibrary's do) — this is a full replace, not a
        partial update. `follow_symlinks`/`show_artists` are the only true
        tri-state fields here: omitted (or explicit null) keeps the
        library's current stored value instead of resetting it — see
        LibraryUpdateRequest.

        Changing `path` wipes and re-scans the library's tracks; the
        response's `path_changed` reports whether that happened, and is
        present only then (omitempty on the handler's side) — never
        present-and-false.

        Refused with 409 `config_managed` for a config-provisioned library
        — edit the Libraries section of config.yaml and restart instead.
      tags: [libraries]
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LibraryUpdateRequest' }
      responses:
        '200':
          description: >
            Updated. The Library shape plus `path_changed` — see the
            operation description.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Library'
                  - type: object
                    properties:
                      path_changed:
                        type: boolean
                        description: >
                          True when this update changed `path`. Omitted
                          entirely — not present-and-false — when the path
                          was unchanged.
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
        '500': { $ref: '#/components/responses/InternalServerError' }
    delete:
      operationId: deleteLibrary
      summary: Delete a library and its tracks.
      description: >
        Admin. Refused with 409 `config_managed` for a config-provisioned
        library, for the same reason as updateLibrary.
      tags: [libraries]
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '204':
          description: Deleted. No body.
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '500': { $ref: '#/components/responses/InternalServerError' }
  /libraries/browse:
    get:
      operationId: browseLibraryPath
      summary: List the immediate subdirectories of an absolute server path (the library path picker).
      description: >
        Admin. Backs the filesystem picker shown when creating or editing a
        library. Unlike `listMetadataFolders`, this is not confined to any
        library root — it has to work before a library exists — so `path`
        here is an arbitrary absolute path on the server filesystem,
        distinct from the metadata-browsing surface's library-relative
        `path` (the shared `LibraryPath` parameter, confined to a library
        root). Symlinked directories are listed and navigable
        (`is_symlink`); the reported path is the one typed/followed, not
        resolved, so pointing a library at a symlink keeps that symlink as
        its stored path. Dot-directories are omitted unless `show_hidden`
        is set.
      tags: [libraries]
      parameters:
        - name: path
          in: query
          required: false
          description: >
            An absolute server filesystem path to list; defaults to "/".
            See the operation description for how this differs from the
            shared `LibraryPath` parameter.
          schema: { type: string }
        - name: show_hidden
          in: query
          required: false
          description: Include dot-directories in the listing. Defaults to false.
          schema: { type: boolean }
      responses:
        '200':
          description: The requested directory's immediate subdirectories, sorted by name.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LibraryBrowseResult' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '500': { $ref: '#/components/responses/InternalServerError' }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: _aether_login
      description: >
        Native-mode session cookie (see docs/agents/authentication.md).
        `/api/v0` defaults every route to admin-only, so a valid, enabled
        admin session is required for the whole metadata surface. This is the
        only credential `/api/v0` actually validates: no route on this
        surface reads an `Authorization` header, and Aether's PATs
        (bearer-shaped credentials) are verified exclusively on `/rest` (via
        `apiKey`, or `u`+`t`+`s` / `u`+`p`), never here.


        In **proxy-header mode**, this cookie plays no role either — identity
        instead comes from reverse-proxy-injected headers (`Remote-User` /
        `Remote-Groups`), validated by `headerGuard`. That trust is
        established between the proxy and Aether, not chosen by the API
        caller, so it has no client-facing credential to model as an OpenAPI
        security scheme; it is intentionally left undeclared here rather than
        represented by a scheme nothing in the protocol actually sends.
  parameters:
...
    ResourceId:
      name: id
      in: path
      required: true
      description: Numeric resource id.
      schema: { type: integer, format: int64, minimum: 0 }
...
  schemas:
    Problem:
      type: object
      description: >
        RFC 9457 Problem Details. `type` is a stable, never-fetched URI whose
        last path segment mirrors the old ad-hoc "code" string.
      properties:
        type:
          type: string
          format: uri
          description: Never dereferenced; only the last path segment is meant to be read.
          example: "https://aether.local/probs/validation_error"
        title:
          type: string
          example: "Validation error"
        status:
          type: integer
          minimum: 100
          maximum: 599
        detail:
          type: string
        instance:
          type: string
          description: The request path.
      required: [type, title, status]
    FieldError:
      type: object
      properties:
        pointer:
          type: string
          description: JSON Pointer (RFC 6901) to the failing field, e.g. "/paths".
        detail:
          type: string
      required: [pointer, detail]
    ValidationProblem:
      description: >
        A Problem for a well-formed but invalid (422) request, itemising
        which fields failed and why. Every 422 on this surface is produced by
        the same code path and carries exactly one field error today, but
        `errors` is not capped at one.
      allOf:
        - $ref: '#/components/schemas/Problem'
        - type: object
          properties:
            errors:
              type: array
              items: { $ref: '#/components/schemas/FieldError' }
...
    Library:
      type: object
      description: >
        A collection/library row, as returned by listLibraries, getLibrary
        and createLibrary; updateLibrary's response is this shape plus
        `path_changed` (modeled inline on that operation, not here).
      properties:
        id: { type: integer, format: int64, minimum: 0 }
        name: { type: string }
        path: { type: string }
        exclude_patterns:
          type: array
          items: { type: string }
          description: Regex patterns matched against each entry name during a scan.
        follow_symlinks:
          type: boolean
          description: >
            Always present here (unlike LibraryCreateRequest/
            LibraryUpdateRequest, where the same field is an optional
            tri-state pointer).
        show_artists:
          type: boolean
          description: >
            Positive-sense wire form of the store's inverted HideArtists
            column. Always present here — see `follow_symlinks`.
        default_view:
          type: string
          enum: [albums, artists]
        icon:
          type: string
          description: A PrimeIcons name without the "pi pi-" prefix, e.g. "folder-open".
        cover_style:
          type: string
          enum: [auto, classic, bauhaus, rings, waves, poster, remix]
          description: >
            "auto" picks a covergen style deterministically per seed; the
            other values pin one specific style.
        source:
          type: string
          enum: [db, config]
          description: >
            "db": created and owned through this API. "config": declared in
            the server config file and materialized here at startup —
            read-only over this API (updateLibrary/deleteLibrary answer 409
            `config_managed`).
        last_scan_started_at:
          type: [string, 'null']
          format: date-time
          description: Present but null until a scan of this library has started.
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        track_count:
          type: integer
          format: int64
          minimum: 0
      required:
        - id
        - name
        - path
        - exclude_patterns
        - follow_symlinks
        - show_artists
        - default_view
        - icon
        - cover_style
        - source
        - last_scan_started_at
        - created_at
        - updated_at
        - track_count
    LibraryList:
      type: object
      properties:
        libraries:
          type: array
          items: { $ref: '#/components/schemas/Library' }
      required: [libraries]
    LibraryCreateRequest:
      type: object
      description: >
        `follow_symlinks`/`show_artists` are optional pointers: omitting
        either here takes the create-time default (both true) — unlike
        LibraryUpdateRequest, where omitting them means "keep current"
        instead. The response always echoes the resolved value back as a
        plain, always-present boolean (see `Library`).
      properties:
        name:
          type: string
          maxLength: 200
          description: Required (400 if blank); over 200 characters is 422.
        path:
          type: string
          description: >
            Required (400 if blank). Resolved to an absolute path
            server-side; must already exist and be a readable directory, or
            this is 422.
        exclude_patterns:
          type: array
          items: { type: string }
          description: >
            Regex patterns matched against each entry name during a scan.
            Omitted defaults to none. Each must compile as a Go regexp (422
            itemising `/exclude_patterns` otherwise).
        follow_symlinks:
          type: boolean
          description: Omitted defaults to true.
        show_artists:
          type: boolean
          description: Omitted defaults to true.
        default_view:
          type: string
          enum: [albums, artists]
          default: albums
          description: Omitted or empty defaults to "albums".
        icon:
          type: string
          pattern: '^[a-z0-9]+(-[a-z0-9]+)*$'
          maxLength: 100
          default: folder
          description: >
            A PrimeIcons name without the "pi pi-" prefix, e.g.
            "folder-open". Omitted or empty defaults to "folder".
        cover_style:
          type: string
          enum: [auto, classic, bauhaus, rings, waves, poster, remix]
          default: auto
          description: Omitted or empty defaults to "auto".
      required: [name, path]
    LibraryUpdateRequest:
      type: object
      description: >
        A full replace, not a partial update: `name`/`path`/
        `exclude_patterns`/`default_view`/`icon`/`cover_style` are always
        overwritten by the submitted value (empty/omitted `default_view`,
        `icon` and `cover_style` fall back the same way
        LibraryCreateRequest's do). `follow_symlinks`/`show_artists` are
        the only true tri-state fields: omitted (or explicit null) keeps
        the library's current stored value instead of resetting it.
        Refused with 409 `config_managed` for a config-provisioned library.
      properties:
        name:
          type: string
          maxLength: 200
          description: Required (400 if blank); over 200 characters is 422.
        path:
          type: string
          description: >
            Required (400 if blank). Resolved to an absolute path
            server-side; must already exist and be a readable directory, or
            this is 422. Changing it wipes and re-scans the library's
            tracks — see updateLibrary's `path_changed` response field.
        exclude_patterns:
          type: array
          items: { type: string }
          description: >
            Replaces the stored patterns entirely; omitted clears them to
            none (this field is not tri-state). Each must compile as a Go
            regexp (422 itemising `/exclude_patterns` otherwise).
        follow_symlinks:
          type: boolean
          description: Omitted (or explicit null) keeps the library's current value.
        show_artists:
          type: boolean
          description: Omitted (or explicit null) keeps the library's current value.
        default_view:
          type: string
          enum: [albums, artists]
          default: albums
          description: Omitted or empty falls back to "albums".
        icon:
          type: string
          pattern: '^[a-z0-9]+(-[a-z0-9]+)*$'
          maxLength: 100
          default: folder
          description: Omitted or empty falls back to "folder".
        cover_style:
          type: string
          enum: [auto, classic, bauhaus, rings, waves, poster, remix]
          default: auto
          description: Omitted or empty falls back to "auto".
      required: [name, path]
    LibraryBrowseResult:
      type: object
      properties:
        path:
          type: string
          description: The requested directory, echoed back cleaned but not symlink-resolved.
        folders:
          type: array
          items: { $ref: '#/components/schemas/BrowseFolder' }
      required: [path, folders]
...
  responses:
    BadRequest:
      description: Malformed request (bad JSON/multipart, missing/invalid scalar).
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    NotFound:
      description: The addressed resource does not exist.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    UnprocessableEntity:
      description: >
        The request is well-formed but semantically invalid (over a size
        cap, an unknown type/slot, an empty selection); `errors` itemises
        the failing fields.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ValidationProblem' }
...
    Unauthorized:
      description: No valid session (native mode) — the caller must authenticate.
      headers:
        WWW-Authenticate:
          schema: { type: string }
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Forbidden:
      description: Authenticated but not permitted (admin required, or the user is disabled).
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Conflict:
      description: The request conflicts with current server state.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
    InternalServerError:
      description: >
        An unexpected server-side failure — a store or internal error that is
        not a not-found (e.g. a failed database read/write, a filesystem write
        that did not complete). The generic Problem shape; see `type`/`detail`.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }